refactor: move thread runtime pool factory
